Conflicts were a duplicate fix. The merge base carried a rogue repo-wide
rename that emitted `#define CMD::NET_OPEN 0x4F` in the plain-C
pico/intellivision tree; this branch fixed it in 18adbf40a as CMD_NET_* /
CMD_FUJI_*, and master fixed it in #1635 as NETCMD_* / FUJICMD_*. Master
is upstream-canonical, so its spelling wins in fujibus.h, fujibus.c and
fujinet.c. pico/o2 and pico/astrocade keep CMD_* in their own headers.

Also drops data/webui/config/fujiversal-{o2,astrocade}.yaml. These were
copies of fujiversal-intv.yaml, which #1621 deleted when it moved the web
UI to BUILD_*.yaml inheritance. Git would have merged them in silently and
build_webui.py's board-first lookup would have used them instead of
BUILD_RS232.yaml, making these the only RS232 boards with a bespoke web UI
and dropping cpm_settings/emulator_settings. Both boards now resolve
through BUILD_RS232 -> base, byte-identically to fujiversal-intv.

+++ lib/bus/adamnet/FujiAdamPacket.cpp
@@ -107,7 +107,8 @@ ByteBuffer FujiAdamPacket::serialize() const
 
     ByteBuffer output;
     output.reserve(totalSize);
-    output.push_back(static_cast<uint8_t>((static_cast<uint8_t>(_type) << 4) | _device));
+    output.push_back(static_cast<uint8_t>((static_cast<uint8_t>(_type) << 4)
+                                          | static_cast<uint8_t>(_device)));
 
     if (hasLen)
     {
